moverクラス追加
